[ETCM-518] Added EthMiningJsonMethodsImplicits and EthBlocksJsonMethodsImplicits
