+++ src/main/scala/io/iohk/ethereum/jsonrpc/EthBlocksJsonMethodsImplicits.scala
@@ -1,0 +1,144 @@
+package io.iohk.ethereum.jsonrpc
+
+import io.iohk.ethereum.jsonrpc.EthBlocksService._
+import io.iohk.ethereum.jsonrpc.JsonRpcError.InvalidParams
+import io.iohk.ethereum.jsonrpc.serialization.{ JsonEncoder, JsonMethodDecoder }
+import io.iohk.ethereum.jsonrpc.serialization.JsonMethodDecoder.NoParamsMethodDecoder
+import org.json4s.Extraction
+import org.json4s.JsonAST.{ JArray, JBool, JField, JString, JValue }
+
+object EthBlocksJsonMethodsImplicits extends JsonMethodsImplicits {
+  implicit val eth_chainId = new NoParamsMethodDecoder(ChainIdRequest()) with JsonEncoder[ChainIdResponse] {
+    def encodeJson(t: ChainIdResponse) = encodeAsHex(t.value)
+  }
+
+  implicit val eth_blockNumber = new NoParamsMethodDecoder(BestBlockNumberRequest())
+    with JsonEncoder[BestBlockNumberResponse] {
+    override def encodeJson(t: BestBlockNumberResponse): JValue = Extraction.decompose(t.bestBlockNumber)
+  }
+
+  implicit val eth_getBlockTransactionCountByHash = new JsonMethodDecoder[TxCountByBlockHashRequest]
+    with JsonEncoder[TxCountByBlockHashResponse] {
+    override def decodeJson(params: Option[JArray]): Either[JsonRpcError, TxCountByBlockHashRequest] =
+      params match {
+        case Some(JArray(JString(input) :: Nil)) =>
+          extractHash(input).map(TxCountByBlockHashRequest)
+        case _ => Left(InvalidParams())
+      }
+
+    override def encodeJson(t: TxCountByBlockHashResponse): JValue =
+      Extraction.decompose(t.txsQuantity.map(BigInt(_)))
+  }
+
+  implicit val eth_getBlockByHash = new JsonMethodDecoder[BlockByBlockHashRequest]
+    with JsonEncoder[BlockByBlockHashResponse] {
+    override def decodeJson(params: Option[JArray]): Either[JsonRpcError, BlockByBlockHashRequest] = {
+      params match {
+        case Some(JArray(JString(blockHash) :: JBool(fullTxs) :: Nil)) =>
+          extractHash(blockHash).map(BlockByBlockHashRequest(_, fullTxs))
+        case _ => Left(InvalidParams())
+      }
+    }
+
+    override def encodeJson(t: BlockByBlockHashResponse): JValue =
+      Extraction.decompose(t.blockResponse)
+  }
+
+  implicit val eth_getBlockByNumber = new JsonMethodDecoder[BlockByNumberRequest]
+    with JsonEncoder[BlockByNumberResponse] {
+    override def decodeJson(params: Option[JArray]): Either[JsonRpcError, BlockByNumberRequest] = {
+      params match {
+        case Some(JArray(blockStr :: JBool(fullTxs) :: Nil)) =>
+          extractBlockParam(blockStr).map(BlockByNumberRequest(_, fullTxs))
+        case _ => Left(InvalidParams())
+      }
+    }
+
+    override def encodeJson(t: BlockByNumberResponse): JValue =
+      Extraction.decompose(t.blockResponse)
+  }
+
+  implicit val eth_getUncleByBlockHashAndIndex = new JsonMethodDecoder[UncleByBlockHashAndIndexRequest]
+    with JsonEncoder[UncleByBlockHashAndIndexResponse] {
+    override def decodeJson(params: Option[JArray]): Either[JsonRpcError, UncleByBlockHashAndIndexRequest] =
+      params match {
+        case Some(JArray(JString(blockHash) :: uncleIndex :: Nil)) =>
+          for {
+            hash <- extractHash(blockHash)
+            uncleBlockIndex <- extractQuantity(uncleIndex)
+          } yield UncleByBlockHashAndIndexRequest(hash, uncleBlockIndex)
+        case _ => Left(InvalidParams())
+      }
+
+    override def encodeJson(t: UncleByBlockHashAndIndexResponse): JValue = {
+      val uncleBlockResponse = Extraction.decompose(t.uncleBlockResponse)
+      uncleBlockResponse.removeField {
+        case JField("transactions", _) => true
+        case _ => false
+      }
+    }
+  }
+
+  implicit val eth_getUncleByBlockNumberAndIndex = new JsonMethodDecoder[UncleByBlockNumberAndIndexRequest]
+    with JsonEncoder[UncleByBlockNumberAndIndexResponse] {
+    override def decodeJson(params: Option[JArray]): Either[JsonRpcError, UncleByBlockNumberAndIndexRequest] =
+      params match {
+        case Some(JArray(blockStr :: uncleIndex :: Nil)) =>
+          for {
+            block <- extractBlockParam(blockStr)
+            uncleBlockIndex <- extractQuantity(uncleIndex)
+          } yield UncleByBlockNumberAndIndexRequest(block, uncleBlockIndex)
+        case _ => Left(InvalidParams())
+      }
+
+    override def encodeJson(t: UncleByBlockNumberAndIndexResponse): JValue = {
+      val uncleBlockResponse = Extraction.decompose(t.uncleBlockResponse)
+      uncleBlockResponse.removeField {
+        case JField("transactions", _) => true
+        case _ => false
+      }
+    }
+  }
+
+  implicit val eth_getUncleCountByBlockNumber = new JsonMethodDecoder[GetUncleCountByBlockNumberRequest]
+    with JsonEncoder[GetUncleCountByBlockNumberResponse] {
+    def decodeJson(params: Option[JArray]): Either[JsonRpcError, GetUncleCountByBlockNumberRequest] =
+      params match {
+        case Some(JArray((blockValue: JValue) :: Nil)) =>
+          for {
+            block <- extractBlockParam(blockValue)
+          } yield GetUncleCountByBlockNumberRequest(block)
+        case _ => Left(InvalidParams())
+      }
+
+    def encodeJson(t: GetUncleCountByBlockNumberResponse): JValue = encodeAsHex(t.result)
+  }
+
+  implicit val eth_getUncleCountByBlockHash = new JsonMethodDecoder[GetUncleCountByBlockHashRequest]
+    with JsonEncoder[GetUncleCountByBlockHashResponse] {
+    def decodeJson(params: Option[JArray]): Either[JsonRpcError, GetUncleCountByBlockHashRequest] =
+      params match {
+        case Some(JArray(JString(hash) :: Nil)) =>
+          for {
+            blockHash <- extractHash(hash)
+          } yield GetUncleCountByBlockHashRequest(blockHash)
+        case _ => Left(InvalidParams())
+      }
+
+    def encodeJson(t: GetUncleCountByBlockHashResponse): JValue = encodeAsHex(t.result)
+  }
+
+  implicit val eth_getBlockTransactionCountByNumber = new JsonMethodDecoder[GetBlockTransactionCountByNumberRequest]
+    with JsonEncoder[GetBlockTransactionCountByNumberResponse] {
+    def decodeJson(params: Option[JArray]): Either[JsonRpcError, GetBlockTransactionCountByNumberRequest] =
+      params match {
+        case Some(JArray((blockValue: JValue) :: Nil)) =>
+          for {
+            block <- extractBlockParam(blockValue)
+          } yield GetBlockTransactionCountByNumberRequest(block)
+        case _ => Left(InvalidParams())
+      }
+
+    def encodeJson(t: GetBlockTransactionCountByNumberResponse): JValue = encodeAsHex(t.result)
+  }
+}
